+++ src/main/java/insane96mcp/progressivebosses/module/wither/feature/MinionFeature.java
@@ -62,10 +62,10 @@ public class MinionFeature extends Feature {
 	public int minionAtDifficulty = 1;
 	public int bonusMinionEveryDifficulty = 1;
 	public int maxSpawned = 6;
-	public int maxAround = 20;
-	public int minCooldown = 300;
-	public int maxCooldown = 600;
-	public double cooldownMultiplierBelowHalfHealth = 0.5d;
+	public int maxAround = 18;
+	public int minCooldown = 400;
+	public int maxCooldown = 700;
+	public double cooldownMultiplierBelowHalfHealth = 0.6d;
 	public double bonusSpeedPerDifficulty = 0.03d;
 	public double magicDamageMultiplier = 3.0d;
 	public boolean killMinionOnWitherDeath = true;